+++ idp-profile-spring/src/test/java/net/shibboleth/idp/profile/spring/relyingparty/metadata/LocalDynamicMetadataProviderParserTest.java
@@ -79,7 +79,7 @@ public class LocalDynamicMetadataProviderParserTest extends AbstractMetadataPars
         Assert.assertNotNull(resolver.getParserPool());
         
         Assert.assertEquals(resolver.getNegativeLookupCacheDuration(), Duration.ofMinutes(10));
-        Assert.assertEquals(resolver.getRefreshDelayFactor(), 0.75f);
+        Assert.assertEquals(resolver.getRefreshDelayFactor().floatValue(), 0.75f);
         Assert.assertEquals(resolver.getMinCacheDuration(), Duration.ofMinutes(10));
         Assert.assertEquals(resolver.getMaxCacheDuration(), Duration.ofHours(8));
         Assert.assertEquals(resolver.getMaxIdleEntityData(), Duration.ofHours(8));
